Logger funciona

+++ fake_persons/views.py
@@ -4,16 +4,21 @@ from mongoengine import connect
 from .forms import PersonForm
 from django.core.files.storage import default_storage
 from django.contrib.auth.decorators import login_required
+import logging
 import os
 
 connect('ssbw', host='mongo')
 
+logger = logging.getLogger(__name__)
+
 def index(request):
+    logger.info("Request: home page")
     error = ""
 
     if request.method == "POST":
         f = PersonForm(request.POST, request.FILES)
         if f.is_valid():
+            logger.info("New Person Form is valid")
             p = f.cleaned_data
 
             u = Person()
@@ -23,9 +28,14 @@ def index(request):
             u.phone = p["phone"]
             u.gender = p["gender"]
 
+            # Check if there is an image on FILES
             img = request.FILES.get('image', False)
-            u.image = "img/" + img.name
-            default_storage.save("fake_persons/static/img/" + img.name, img)
+            if img:
+                logger.info("Downloading image for the new person")
+                u.image = "img/" + img.name
+                default_storage.save("fake_persons/static/img/" + img.name, img)
+            else:
+                logger.info("Creating person without image")
 
             a = Address()
             a.street = p["street"]
@@ -40,24 +50,33 @@ def index(request):
             u.save()
         else:
             error = "Couldn't create person"
+            logger.error("Invalid Person Form: %s", error)
 
     persons = Person.objects.all()
     return render(request, 'fake_persons/index.html', {'persons': persons, 'error': error})
 
 @login_required
 def person_detail(request, pk):
+    logger.info("Request: Person detail with id %s", pk)
     p = Person.objects.get(pk=pk)
     return render(request, 'fake_persons/person.html', {'person': p})
 
 @login_required
 def person_new(request):
+    logger.info("Request: New person form")
     f = PersonForm()
     return render(request, 'fake_persons/person_new.html', {'form': f})
 
 @login_required
 def person_delete(request, pk):
+    logger.info("Request: Delete person with id %s", pk)
     if request.method == "POST":
+        logger.info("Person with id ", pk, " deleted")
         p = Person.objects.get(pk=pk)
+
+        # Delete image file
+        os.remove("fake_persons/static/" + p["image"])
+
         p.delete()
 
     persons = Person.objects.all()
@@ -65,6 +84,7 @@ def person_delete(request, pk):
 
 @login_required
 def person_edit(request, pk):
+    logger.info("Request: Edit person with id %s", pk)
     person = Person.objects.get(pk=pk)
     f = PersonForm()
     
@@ -72,8 +92,8 @@ def person_edit(request, pk):
         return render(request, 'fake_persons/person_edit.html', {'form': f, 'person': person})
     elif request.method == "POST":
         f = PersonForm(request.POST, request.FILES)
-        print(f)
         if f.is_valid():
+            logger.info("Edit Person Form is valid")
             p = f.cleaned_data
 
             person.update(set__firstName=p["firstName"])
@@ -82,13 +102,19 @@ def person_edit(request, pk):
             person.update(set__phone=p["phone"])
             person.update(set__gender=p["gender"])
 
-            # Delete old image
-            os.remove("fake_persons/static/" + person["image"])
-
-            # Update new image
+            # Check if there is an image on FILES
             img = request.FILES.get('image', False)
-            person.update(set__image="img/" + img.name)
-            default_storage.save("fake_persons/static/img/" + img.name, img)
+            if img:
+                logger.info("Changing image of person with id %s", pk)
+
+                # Delete old image
+                os.remove("fake_persons/static/" + person["image"])
+
+                # Update new image
+                person.update(set__image="img/" + img.name)
+                default_storage.save("fake_persons/static/img/" + img.name, img)
+            else:
+                logger.info("Keeping image of person with id %s", pk)
 
             a = Address()
             a.street = p["street"]
@@ -99,11 +125,8 @@ def person_edit(request, pk):
             a.country = p["country"]
             a.countryCode = p["countryCode"]
             person.update(set__address=a)
+        else:
+            logger.error("Invalid Edit Person Form")
 
         person = Person.objects.get(pk=pk)
         return render(request, 'fake_persons/person.html', {'person': person})
-
-def download_image(f):
-    with open('img/' + str(f), 'wb+') as destination:
-        for chunk in f.chunks():
-            destination.write(chunk)
